+++ sympy/polys/numberfields/tests/test_numbers.py
@@ -1,36 +1,15 @@
 """Tests on algebraic numbers. """
 
 from sympy.core.containers import Tuple
-from sympy.core.numbers import (AlgebraicNumber, I, Rational)
+from sympy.core.numbers import (AlgebraicNumber, Rational)
 from sympy.core.singleton import S
 from sympy.core.symbol import Symbol
 from sympy.functions.elementary.miscellaneous import sqrt
 from sympy.polys.polytools import Poly
-from sympy.testing.pytest import raises
-from sympy.polys.numberfields.numbers import (
-    to_number_field,
-    isolate,
-)
-from sympy.polys.polyerrors import IsomorphismFailed
+from sympy.polys.numberfields.subfield import to_number_field
 from sympy.polys.polyclasses import DMP
 from sympy.polys.domains import QQ
-from sympy.printing.lambdarepr import IntervalPrinter
 from sympy.abc import x, y
-
-Q = Rational
-
-
-def test_to_number_field():
-    assert to_number_field(sqrt(2)) == AlgebraicNumber(sqrt(2))
-    assert to_number_field(
-        [sqrt(2), sqrt(3)]) == AlgebraicNumber(sqrt(2) + sqrt(3))
-
-    a = AlgebraicNumber(sqrt(2) + sqrt(3), [S.Half, S.Zero, Rational(-9, 2), S.Zero])
-
-    assert to_number_field(sqrt(2), sqrt(2) + sqrt(3)) == a
-    assert to_number_field(sqrt(2), AlgebraicNumber(sqrt(2) + sqrt(3))) == a
-
-    raises(IsomorphismFailed, lambda: to_number_field(sqrt(2), sqrt(3)))
 
 
 def test_AlgebraicNumber():
@@ -194,22 +173,3 @@ def test_to_algebraic_integer():
     assert a.minpoly == x**2 - 12
     assert a.root == 2*sqrt(3)
     assert a.rep == DMP([QQ(7, 19), QQ(3)], QQ)
-
-
-def test_IntervalPrinter():
-    ip = IntervalPrinter()
-    assert ip.doprint(x**Q(1, 3)) == "x**(mpi('1/3'))"
-    assert ip.doprint(sqrt(x)) == "x**(mpi('1/2'))"
-
-
-def test_isolate():
-    assert isolate(1) == (1, 1)
-    assert isolate(S.Half) == (S.Half, S.Half)
-
-    assert isolate(sqrt(2)) == (1, 2)
-    assert isolate(-sqrt(2)) == (-2, -1)
-
-    assert isolate(sqrt(2), eps=Rational(1, 100)) == (Rational(24, 17), Rational(17, 12))
-    assert isolate(-sqrt(2), eps=Rational(1, 100)) == (Rational(-17, 12), Rational(-24, 17))
-
-    raises(NotImplementedError, lambda: isolate(I))
