+++ sympy/polys/numberfields/utilities.py
@@ -1,0 +1,45 @@
+"""Utilities for algebraic number theory. """
+
+from sympy.core.sympify import sympify
+from sympy.polys.numberfields.minpoly import minpoly
+from sympy.printing.lambdarepr import IntervalPrinter
+from sympy.utilities import lambdify, public
+
+from mpmath import mp
+
+
+@public
+def isolate(alg, eps=None, fast=False):
+    """Give a rational isolating interval for an algebraic number. """
+    alg = sympify(alg)
+
+    if alg.is_Rational:
+        return (alg, alg)
+    elif not alg.is_real:
+        raise NotImplementedError(
+            "complex algebraic numbers are not supported")
+
+    func = lambdify((), alg, modules="mpmath", printer=IntervalPrinter())
+
+    poly = minpoly(alg, polys=True)
+    intervals = poly.intervals(sqf=True)
+
+    dps, done = mp.dps, False
+
+    try:
+        while not done:
+            alg = func()
+
+            for a, b in intervals:
+                if a <= alg.a and alg.b <= b:
+                    done = True
+                    break
+            else:
+                mp.dps *= 2
+    finally:
+        mp.dps = dps
+
+    if eps is not None:
+        a, b = poly.refine_root(a, b, eps=eps, fast=fast)
+
+    return (a, b)
